mannual pick from master: [improvement](storage) For debugging problems: add session variable skip_storage_engine_merge to treat agg and unique data model as dup model (#11952)
For debug purpose:
Add session variable skip_storage_engine_merge, when set to true, tables
of aggregate key model and unique key model will be read as duplicate
key model.
Add session variable skip_delete_predicate, when set to true, rows
deleted with delete statement will be selected.

+++ be/src/olap/reader.cpp
@@ -823,8 +823,8 @@ OLAPStatus TabletReader::_init_delete_condition(const ReaderParams& read_params)
     if (read_params.reader_type == READER_CUMULATIVE_COMPACTION) {
         return OLAP_SUCCESS;
     }
-    OLAPStatus ret;
-    {
+    OLAPStatus ret = OLAP_SUCCESS;
+    if (read_params.runtime_state && !read_params.runtime_state->skip_delete_predicate()) {
         ReadLock rdlock(_tablet->get_header_lock());
         ret = _delete_handler.init(_tablet->tablet_schema(), _tablet->delete_predicates(),
                                               read_params.version.second, this);
